CHECK THIS ONE FOR CORRECTNESS. I was not entirely sure of the \= with the section where it says, "*c = ( *a /= len ).cross( *b /= len );". In the case that \= returns a new Point, this code is good. If not, you will need to normalize a and divide b by the length of a (which is a little weird).
The reason I did this is because this is bad and fragile code, and as you can tell by the above, somewhat hard to follow.

Secondly, the compiler can sometimes have a hard time with ailiasing pointers (the things you are creating by your reinterpert casts). In the worst case, it forces the compiler to maintain the memory layout of the objects, which can prevent inlineing and efficent register allocation. Doing it this way costs very little, makes it easier to read, and may give a slight performance improvement.