Manny, I suggest using a static context over your current method of initializing and deinitializing the library on every function invocation.
Currently, you are opening at least one file hande, parsing it, setting up a lot of dynamic memory, extracting one value, and tearing it all down. Its a little bit like building a bakery per cupcake you make.

With the context pattern, the library is opened once in the object constructor when the function is first invoked, and cached for the lifetime of the program (static storage duration). Then when the program is done, the library is automatically closed by the context's destructor.

The context pattern is fine for the one usage here, however if you were to do more subroutines, it would be better to initialize the library in main on startup.